done category

+++ Admin/src/main/java/com/example/admin/Entity/ProductEntity.java
@@ -30,7 +30,7 @@ public class ProductEntity {
 
     private int overall_rating;
 
-    private Long discount;
+    private double discount;
 
     private String image1;
